+++ js/app/coopcycle-frontend-js/logistics/redux.js
@@ -1,0 +1,5 @@
+// published version
+export * from 'coopcycle-frontend-js/logistics/redux'
+
+// local version (to be used only during development)
+// export * from '../../shared-js-lib-local/logistics/redux'
