+++ js/app/dashboard/components/Task.js
@@ -7,7 +7,7 @@ import _ from 'lodash'
 
 import { setCurrentTask, toggleTask, selectTask } from '../redux/actions'
 import { selectIsVisibleTask } from '../redux/selectors'
-import { selectSelectedDate, selectTasksWithColor } from '../../coopcycle-frontend-js/lastmile/redux'
+import { selectSelectedDate, selectTasksWithColor } from '../../coopcycle-frontend-js/logistics/redux'
 
 import { addressAsText } from '../utils'
 import TaskEta from './TaskEta'
