+++ js/app/dashboard/components/TaskLists.js
@@ -8,7 +8,7 @@ import { createTaskList, closeAddUserModal, openAddUserModal, openNewTaskModal, 
 import CourierSelect from './CourierSelect'
 import TaskList from './TaskList'
 
-import { selectTaskLists, selectSelectedDate } from '../../coopcycle-frontend-js/lastmile/redux'
+import { selectTaskLists, selectSelectedDate } from '../../coopcycle-frontend-js/logistics/redux'
 
 class TaskLists extends React.Component {
 
@@ -112,7 +112,7 @@ function mapStateToProps (state) {
     addModalIsOpen: state.addModalIsOpen,
     taskLists: selectTaskLists(state),
     date: selectSelectedDate(state),
-    taskListsLoading: state.lastmile.ui.taskListsLoading,
+    taskListsLoading: state.logistics.ui.taskListsLoading,
     taskModalIsOpen: state.taskModalIsOpen,
   }
 }
