+++ js/app/dashboard/components/TaskModalContent.js
@@ -16,7 +16,7 @@ import CourierSelect from './CourierSelect'
 import { timePickerProps } from '../../utils/antd'
 
 import { closeNewTaskModal, createTask, startTask, completeTask, cancelTask, duplicateTask, loadTaskEvents } from '../redux/actions'
-import { selectSelectedDate } from '../../coopcycle-frontend-js/lastmile/redux'
+import { selectSelectedDate } from '../../coopcycle-frontend-js/logistics/redux'
 
 const itemColor = event => {
   switch (event.name) {
