+++ js/app/dashboard/components/UnassignedTasks.js
@@ -10,7 +10,7 @@ import Task from './Task'
 import TaskGroup from './TaskGroup'
 import UnassignedTasksPopoverContent from './UnassignedTasksPopoverContent'
 import { setTaskListGroupMode, openNewTaskModal, closeNewTaskModal, toggleSearch } from '../redux/actions'
-import { selectUnassignedTasks } from '../../coopcycle-frontend-js/lastmile/redux'
+import { selectUnassignedTasks } from '../../coopcycle-frontend-js/logistics/redux'
 
 const UnassignedTasksPopoverContentWithTrans = withTranslation()(UnassignedTasksPopoverContent)
