+++ js/app/dashboard/redux/middlewares.js
@@ -86,7 +86,7 @@ export const socketIO = ({ dispatch, getState }) => {
 }
 
 function getKey(state) {
-  return state.lastmile.date.format('YYYY-MM-DD')
+  return state.logistics.date.format('YYYY-MM-DD')
 }
 
 export const persistFilters = ({ getState }) => (next) => (action) => {
